Add initial implementation of project assignment
Added initial implementation of the project assignment. This includes
a Table of Contents, functional requirements, and a list of resources
such as datasets and contacts. A frontend has been implemented to
display the journey list, station list, and single station view using
the Pyramid web application framework. The implementation is
containerized using Cloud run. Overall, this commit adds 364 lines of
code.