Update dataset exploration section and parser
Updated the section title in README.org to correct a typo, from
"Exlpore" to "Explore", and made some changes to the parser to handle
the dataset more robustly. Distance is now set to 0 if it's
represented as an empty string, and stations names and ids are checked
to be strings. Also added a parser class with an example usage for the
three CSV files in the dataset to check if all data is parseable.