Add implementation for journeys list view and database connection.
This commit adds the implementation for the journeys list view, along
with the configuration of the sqlite database backend in
hcbf/development.ini and hcbf/hcbf/__init__.py. It also includes a new
function `db_connect` which creates a new database connection on each
request and adds a callback to close the connection once the request
has finished. The `db_connect` function can then be used as a request
method to retrieve a database connection in a view function.