Add frontend for journey list view
The commit adds the frontend for the journey list view to the
codebase. It makes use of a Flask app and SQLite rdbms to retrieve
journeys from the database and renders the result as an HTML table on
the frontend. The commit also includes a python class responsible for
parsing journey records.