Add functionality to export data to a PostgreSQL server
In this commit, I added a new feature to export the data from the
'journey' table in the SQLite database to a PostgreSQL server. I
created a new table named 'journey' in the PostgreSQL database and
used psycopg2 to connect to the database. I also created an index on
the table to perform stable sorting. Finally, I closed the connections
to both the databases.