Add support for PostgreSQL as a data source in the frontend
In this commit, I added support for PostgreSQL as a data source to the
frontend of the application. I created an abstract base class for the
database, implemented the SQLite and PostgreSQL databases as concrete
classes, and modified the `journeys()` method to use the appropriate
database. I also added `psycopg2-binary` to the dependencies list.