Add new parsers for journeys and station datasets
This commit adds new parsers for the journeys and station datasets in
`README.org`. The journeys parser contains minor code to merge the
files into a single file, remove duplicates, and update the statistics
on the merged file. Meanwhile, the station parser is implemented as a
single script that imports the requirements from the
`station-to-sql-requirements.txt` file and connects to the PostgreSQL
database to create the station table and insert the records.

The station record is defined using `namedtuple`, which is then used
to set the properties of an instance of `Station`. Specifically, the
properties are fid, id, finnish_name, swedish_name, english_name,
finnish_address, swedish_address, finnish_city, swedish_city,
operator, capacity, x, and y. Each property is validated according to
the field's datatype (e.g., `capacity` is validated to be integer).