Add implementation for getting station details along with counts
The commit implements the `get_station_with_departures_and_arrivals`
method in the `PostgreDB` class to get the station details along with
the number of journeys starting and ending there. The data returned by
this method is useful for rendering the Station page.

The commit also adds new named tuple `StationRecordPlus` and updates
the Station view to use the count of arrivals and departures, which
are part of the new named tuple.

Lastly, the commit adds a fix to the size of the map on the Station
page, set to a minimum height of 400px.