Implement filtering by month in single station view
This commit updates the `get_station_statistics` method in `PostgreDB`
to take a `months` parameter and filter journeys by the given
months. It also adds a new form in the `station.html.j2` template that
allows the user to select which months to filter by. This will display
statistics based on journeys occurring in the selected months.