Add functionality to redirect user to a random station
This commit adds a new route `/random-station` which queries the
PostgreSQL database for a random station ID and redirects the user to
the `station` page for that station. The functionality was implemented
in `main.py` using the `get_random_station_id` method of the
`PostgreDB` class, and a new link was added to the `home.html.j2`
template for users to pick a random station.

Also, a minor modification was made to the `README.org` file to
include the new SQL code block for selecting a random station ID.

Overall, this commit adds more fun and interactivity to the app,
allowing users to explore the stats for a random station without
knowing its ID beforehand.