Add implementation for backend REST API for getting journeys
This commit adds implementation for getting journeys in the backend of
the app. It includes a new endpoint `/journeys` that accepts query
parameters for pagination, search, ordering and direction. The
implementation uses `psycopg2` library for executing SQL queries with
PostgreSQL. The result is serialized using `flask_restful` package and
returned as JSON.