Add implementation for frontend using backend REST API
This commit updates the app to use the newly implemented `/journeys`
endpoint of the backend service for rendering journeys in the
frontend. The implementation uses the `requests` library to fetch the
data and render it using Jinja2 templates. The query parameters are
dynamically passed to the backends through the URL, and the result is
deserialized in the frontend as a list of Journey dataclasses. The
journeys are then rendered on a table in the corresponding HTML
template. The previous and next pages are computed based on the
current page.