Refactor API endpoint to return random station ID
In this commit, we refactor the API endpoint for getting a random
station ID to be handled by a new resource called
"RandomStation". This resource simply returns a randomly selected
station ID from the database using the "get_random_station_id" method
of the DB class. We also update the Flask app to use this new resource
and modify the "random_station" route handler to simply redirect to
the "station" route with the randomly selected station ID. This makes
the API more modular and easier to maintain, while also simplifying
our code for generating a random station ID.